Refactored GHashCommon
